// workspaces/sykmeldinger-next/src/server/graphql/resolver-types.generated.ts
export type Maybe<T> = T | null
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> }
// ...
export type Brukerinformasjon = {
    __typename?: 'Brukerinformasjon'
    arbeidsgivere: Array<Arbeidsgiver>
}
// ...
export type Mutation = {
    __typename?: 'Mutation'
    changeSykmeldingStatus: Sykmelding
    dev_changeScenario: Scalars['Boolean']['output']
    dev_setAntallArbeidsgivere: Scalars['Boolean']['output']
    feedback: Scalars['Boolean']['output']
    sendSykmelding: Sykmelding
    updateEgenmeldingsdager: Sykmelding
}

export type MutationChangeSykmeldingStatusArgs = {
    status: SykmeldingChangeStatus
    sykmeldingId: Scalars['String']['input']
}

export type MutationDev_ChangeScenarioArgs = {
    scenario: Scalars['String']['input']
}

export type MutationDev_SetAntallArbeidsgivereArgs = {
    antall: Scalars['Int']['input']
}

export type MutationFeedbackArgs = {
    feedback: Scalars['JSON']['input']
}

export type MutationSendSykmeldingArgs = {
    sykmeldingId: Scalars['String']['input']
    values: SendSykmeldingValues
}

export type MutationUpdateEgenmeldingsdagerArgs = {
    egenmeldingsdager: Array<Scalars['Date']['input']>
    sykmeldingId: Scalars['String']['input']
}
// ...
export type Query = {
    __typename?: 'Query'
    brukerinformasjon: Brukerinformasjon
    sykmelding: Sykmelding
    sykmeldingUtenforVentetid: UtenforVentetid
    sykmeldinger: Array<Sykmelding>
    tidligereArbeidsgivere: Maybe<Array<TidligereArbeidsgiver>>
}

export type QuerySykmeldingArgs = {
    id: Scalars['String']['input']
}
// ...
export type Sykmelding = {
    __typename?: 'Sykmelding'
    andreTiltak: Maybe<Scalars['String']['output']>
    arbeidsgiver: Maybe<ArbeidsgiverSykmelding>
    behandler: Behandler
    behandletTidspunkt: Scalars['Date']['output']
    behandlingsutfall: Behandlingsutfall
    egenmeldt: Maybe<Scalars['Boolean']['output']>
    id: Scalars['String']['output']
    kontaktMedPasient: KontaktMedPasient
    medisinskVurdering: Maybe<MedisinskVurdering>
    meldingTilArbeidsgiver: Maybe<Scalars['String']['output']>
    meldingTilNAV: Maybe<MeldingTilNav>
    merknader: Maybe<Array<Merknad>>
    mottattTidspunkt: Scalars['Date']['output']
    papirsykmelding: Maybe<Scalars['Boolean']['output']>
    pasient: Maybe<Pasient>
    prognose: Maybe<Prognose>
    rulesetVersion: Scalars['Int']['output']
    sykmeldingStatus: SykmeldingStatus
    sykmeldingsperioder: Array<Periode>
    tiltakArbeidsplassen: Maybe<Scalars['String']['output']>
    tiltakNAV: Maybe<Scalars['String']['output']>
    utdypendeOpplysninger: Scalars['JSON']['output']
    utenlandskSykmelding: Maybe<UtenlandskSykmelding>
}
// ...
export type UtenforVentetid = {
    __typename?: 'UtenforVentetid'
    erUtenforVentetid: Scalars['Boolean']['output']
    oppfolgingsdato: Maybe<Scalars['Date']['output']>
}

export type UtenlandskSykmelding = {
    __typename?: 'UtenlandskSykmelding'
    land: Scalars['String']['output']
}

export enum YesOrNo {
    NO = 'NO',
    YES = 'YES',
}

export type WithIndex<TObject> = TObject & Record<string, any>
export type ResolversObject<TObject> = WithIndex<TObject>

export type ResolverTypeWrapper<T> = Promise<T> | T

export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>
}
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
    | ResolverFn<TResult, TParent, TContext, TArgs>
    | ResolverWithResolve<TResult, TParent, TContext, TArgs>
// ...
export type ResolversParentTypes = ResolversObject<{
    Adresse: Adresse
    AktivitetIkkeMuligPeriode: AktivitetIkkeMuligPeriode
    AnnenFraversArsak: AnnenFraversArsak
    Arbeidsgiver: Arbeidsgiver
    ArbeidsgiverOrgnummerBrukerSvar: ArbeidsgiverOrgnummerBrukerSvar
    ArbeidsgiverStatus: ArbeidsgiverStatus
    ArbeidsgiverSykmelding: ArbeidsgiverSykmelding
    ArbeidsledigBrukerSvar: ArbeidsledigBrukerSvar
    ArbeidsledigFraOrgnummerBrukerSvar: ArbeidsledigFraOrgnummerBrukerSvar
    ArbeidsledigInput: ArbeidsledigInput
    ArbeidsrelatertArsak: ArbeidsrelatertArsak
    ArbeidssituasjonBrukerSvar: ArbeidssituasjonBrukerSvar
    ArbeidssituasjonSvar: ArbeidssituasjonSvar
    Behandler: Behandler
    Behandlingsutfall: Behandlingsutfall
    BladBrukerSvar: BladBrukerSvar
    Boolean: Scalars['Boolean']['output']
    BrukerSvar: BrukerSvar
    Brukerinformasjon: Brukerinformasjon
    DagerSvar: DagerSvar
    Date: Scalars['Date']['output']
    DateRange: DateRange
    DateTime: Scalars['DateTime']['output']
    Diagnose: Diagnose
    EgenmeldingsdagerBrukerSvar: EgenmeldingsdagerBrukerSvar
    ErIArbeid: ErIArbeid
    ErIkkeIArbeid: ErIkkeIArbeid
    ErOpplysningeneRiktigeBrukerSvar: ErOpplysningeneRiktigeBrukerSvar
    FiskerBrukerSvar: FiskerBrukerSvar
    FiskerInput: FiskerInput
    FomTom: FomTom
    FrilanserEllerSelvstendigEgenmeldingsperioderBrukerSvar: FrilanserEllerSelvstendigEgenmeldingsperioderBrukerSvar
    GradertPeriode: GradertPeriode
    HarBruktEgenmeldingsdagerBrukerSvar: HarBruktEgenmeldingsdagerBrukerSvar
    HarForsikringBrukerSvar: HarForsikringBrukerSvar
    HarFrilanserEllerSelvstendigBruktEgenmeldingBrukerSvar: HarFrilanserEllerSelvstendigBruktEgenmeldingBrukerSvar
    Int: Scalars['Int']['output']
    JSON: Scalars['JSON']['output']
    JaNeiSvar: JaNeiSvar
    KontaktMedPasient: KontaktMedPasient
    LottOgHyreBrukerSvar: LottOgHyreBrukerSvar
    MedisinskArsak: MedisinskArsak
    MedisinskVurdering: MedisinskVurdering
    MeldingTilNAV: MeldingTilNav
    Merknad: Merknad
    Mutation: {}
    NaermesteLeder: NaermesteLeder
    Pasient: Pasient
    Periode: Periode
    PerioderSvar: PerioderSvar
    Prognose: Prognose
    Query: {}
    RegelInfo: RegelInfo
    RiktigNarmesteLederBrukerSvar: RiktigNarmesteLederBrukerSvar
    SendSykmeldingValues: SendSykmeldingValues
    Sporsmal: Omit<Sporsmal, 'svar'> & { svar: ResolversParentTypes['SvarTypeUnion'] }
    String: Scalars['String']['output']
    SvarTypeUnion: ResolversUnionTypes<ResolversParentTypes>['SvarTypeUnion']
    Sykmelding: Sykmelding
    SykmeldingStatus: SykmeldingStatus
    TidligereArbeidsgiver: TidligereArbeidsgiver
    UriktigeOpplysningerBrukerSvar: UriktigeOpplysningerBrukerSvar
    UtdypendeOpplysning: UtdypendeOpplysning
    UtenforVentetid: UtenforVentetid
    UtenlandskSykmelding: UtenlandskSykmelding
}>
// ...
export type MutationResolvers<
    ContextType = RequestContext,
    ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation'],
> = ResolversObject<{
    changeSykmeldingStatus?: Resolver<
        ResolversTypes['Sykmelding'],
        ParentType,
        ContextType,
        RequireFields<MutationChangeSykmeldingStatusArgs, 'status' | 'sykmeldingId'>
    >
    dev_changeScenario?: Resolver<
        ResolversTypes['Boolean'],
        ParentType,
        ContextType,
        RequireFields<MutationDev_ChangeScenarioArgs, 'scenario'>
    >
    dev_setAntallArbeidsgivere?: Resolver<
        ResolversTypes['Boolean'],
        ParentType,
        ContextType,
        RequireFields<MutationDev_SetAntallArbeidsgivereArgs, 'antall'>
    >
    feedback?: Resolver<
        ResolversTypes['Boolean'],
        ParentType,
        ContextType,
        RequireFields<MutationFeedbackArgs, 'feedback'>
    >
    sendSykmelding?: Resolver<
        ResolversTypes['Sykmelding'],
        ParentType,
        ContextType,
        RequireFields<MutationSendSykmeldingArgs, 'sykmeldingId' | 'values'>
    >
    updateEgenmeldingsdager?: Resolver<
        ResolversTypes['Sykmelding'],
        ParentType,
        ContextType,
        RequireFields<MutationUpdateEgenmeldingsdagerArgs, 'egenmeldingsdager' | 'sykmeldingId'>
    >
}>
// ...
export type QueryResolvers<
    ContextType = RequestContext,
    ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query'],
> = ResolversObject<{
    brukerinformasjon?: Resolver<ResolversTypes['Brukerinformasjon'], ParentType, ContextType>
    sykmelding?: Resolver<
        ResolversTypes['Sykmelding'],
        ParentType,
        ContextType,
        RequireFields<QuerySykmeldingArgs, 'id'>
    >
    sykmeldingUtenforVentetid?: Resolver<
        ResolversTypes['UtenforVentetid'],
        ParentType,
        ContextType,
        RequireFields<QuerySykmeldingUtenforVentetidArgs, 'id'>
    >
    sykmeldinger?: Resolver<Array<ResolversTypes['Sykmelding']>, ParentType, ContextType>
    tidligereArbeidsgivere?: Resolver<
        Maybe<Array<ResolversTypes['TidligereArbeidsgiver']>>,
        ParentType,
        ContextType,
        RequireFields<QueryTidligereArbeidsgivereArgs, 'id'>
    >
}>
// ...
export type Resolvers<ContextType = RequestContext> = ResolversObject<{
    Adresse?: AdresseResolvers<ContextType>
    AktivitetIkkeMuligPeriode?: AktivitetIkkeMuligPeriodeResolvers<ContextType>
    AnnenFraversArsak?: AnnenFraversArsakResolvers<ContextType>
    Arbeidsgiver?: ArbeidsgiverResolvers<ContextType>
    ArbeidsgiverOrgnummerBrukerSvar?: ArbeidsgiverOrgnummerBrukerSvarResolvers<ContextType>
    ArbeidsgiverStatus?: ArbeidsgiverStatusResolvers<ContextType>
    ArbeidsgiverSykmelding?: ArbeidsgiverSykmeldingResolvers<ContextType>
    ArbeidsledigBrukerSvar?: ArbeidsledigBrukerSvarResolvers<ContextType>
    ArbeidsledigFraOrgnummerBrukerSvar?: ArbeidsledigFraOrgnummerBrukerSvarResolvers<ContextType>
    ArbeidsrelatertArsak?: ArbeidsrelatertArsakResolvers<ContextType>
    ArbeidssituasjonBrukerSvar?: ArbeidssituasjonBrukerSvarResolvers<ContextType>
    ArbeidssituasjonSvar?: ArbeidssituasjonSvarResolvers<ContextType>
    Behandler?: BehandlerResolvers<ContextType>
    Behandlingsutfall?: BehandlingsutfallResolvers<ContextType>
    BladBrukerSvar?: BladBrukerSvarResolvers<ContextType>
    BrukerSvar?: BrukerSvarResolvers<ContextType>
    Brukerinformasjon?: BrukerinformasjonResolvers<ContextType>
    DagerSvar?: DagerSvarResolvers<ContextType>
    Date?: GraphQLScalarType
    DateTime?: GraphQLScalarType
    Diagnose?: DiagnoseResolvers<ContextType>
    EgenmeldingsdagerBrukerSvar?: EgenmeldingsdagerBrukerSvarResolvers<ContextType>
    ErIArbeid?: ErIArbeidResolvers<ContextType>
    ErIkkeIArbeid?: ErIkkeIArbeidResolvers<ContextType>
    ErOpplysningeneRiktigeBrukerSvar?: ErOpplysningeneRiktigeBrukerSvarResolvers<ContextType>
    FiskerBrukerSvar?: FiskerBrukerSvarResolvers<ContextType>
    FomTom?: FomTomResolvers<ContextType>
    FrilanserEllerSelvstendigEgenmeldingsperioderBrukerSvar?: FrilanserEllerSelvstendigEgenmeldingsperioderBrukerSvarResolvers<ContextType>
    GradertPeriode?: GradertPeriodeResolvers<ContextType>
    HarBruktEgenmeldingsdagerBrukerSvar?: HarBruktEgenmeldingsdagerBrukerSvarResolvers<ContextType>
    HarForsikringBrukerSvar?: HarForsikringBrukerSvarResolvers<ContextType>
    HarFrilanserEllerSelvstendigBruktEgenmeldingBrukerSvar?: HarFrilanserEllerSelvstendigBruktEgenmeldingBrukerSvarResolvers<ContextType>
    JSON?: GraphQLScalarType
    JaNeiSvar?: JaNeiSvarResolvers<ContextType>
    KontaktMedPasient?: KontaktMedPasientResolvers<ContextType>
    LottOgHyreBrukerSvar?: LottOgHyreBrukerSvarResolvers<ContextType>
    MedisinskArsak?: MedisinskArsakResolvers<ContextType>
    MedisinskVurdering?: MedisinskVurderingResolvers<ContextType>
    MeldingTilNAV?: MeldingTilNavResolvers<ContextType>
    Merknad?: MerknadResolvers<ContextType>
    Mutation?: MutationResolvers<ContextType>
    NaermesteLeder?: NaermesteLederResolvers<ContextType>
    Pasient?: PasientResolvers<ContextType>
    Periode?: PeriodeResolvers<ContextType>
    PerioderSvar?: PerioderSvarResolvers<ContextType>
    Prognose?: PrognoseResolvers<ContextType>
    Query?: QueryResolvers<ContextType>
    RegelInfo?: RegelInfoResolvers<ContextType>
    RiktigNarmesteLederBrukerSvar?: RiktigNarmesteLederBrukerSvarResolvers<ContextType>
    Sporsmal?: SporsmalResolvers<ContextType>
    SvarTypeUnion?: SvarTypeUnionResolvers<ContextType>
    Sykmelding?: SykmeldingResolvers<ContextType>
    SykmeldingStatus?: SykmeldingStatusResolvers<ContextType>
    TidligereArbeidsgiver?: TidligereArbeidsgiverResolvers<ContextType>
    UriktigeOpplysningerBrukerSvar?: UriktigeOpplysningerBrukerSvarResolvers<ContextType>
    UtdypendeOpplysning?: UtdypendeOpplysningResolvers<ContextType>
    UtenforVentetid?: UtenforVentetidResolvers<ContextType>
    UtenlandskSykmelding?: UtenlandskSykmeldingResolvers<ContextType>
}>